more wrap work
moved to solution that impliments two new treeblocks, "WrapGap" and "WrapLine". this alows the parser to not consume these blocks ahead of time

currently the program functions and is navigable, but is not saving updated content on keypress. inserting and deleting character(s) is not possible because preLineWrap() does not read data from .raw.ref.value, only from .exe.ref.textContent! This results in the user's typing nevery actually being read prior to parsing.