checklists + fix `Redir`s
impliment system for formatting checklists ([Y] -> [✓], [N] -> [✗], [~] -> [~]).

also finished updating the ./Redir/.* to be .html.php so it can use the clean php import system for the page header metadata. this was made necessary by the addition of the checklist syntax to the Text Formatting guide (`./Redir/textformat.html.php`)