resolve #3
line 491 of `main.js` was improperly constituting HTML elements blindly without properly escaping HTML special chars. by using `Formatter.escapeHTML` on it, there is no chance to do this.

The reason the issue was only happening if you pressed enter AFTER the `<tag>`, was because it is evaluating `textarea.value.substring(0, textarea.selectionEnd)`. This makese sense